rellocate code and add tests
